preserve server.go git history
